Feat: add cypress settings, update folders & commands
