+++ cypress/.eslintrc.js
@@ -1,4 +1,5 @@
 module.exports = {
+  root: true,
   plugins: ['eslint-plugin-cypress'],
   extends: ['plugin:cypress/recommended'],
   env: {
